+++ src/Mapping/Event/Adapter/ODM.php
@@ -150,6 +150,8 @@ class ODM implements AdapterInterface
     }
 
     /**
+     * @deprecated to be removed in 4.0, use custom lifecycle event classes instead.
+     *
      * Creates a ODM specific LifecycleEventArgs.
      *
      * @param object          $document
@@ -159,6 +161,11 @@ class ODM implements AdapterInterface
      */
     public function createLifecycleEventArgsInstance($document, $documentManager)
     {
+        @trigger_error(sprintf(
+            'Using "%s()" method is deprecated since gedmo/doctrine-extensions 3.15 and will be removed in version 4.0.',
+            __METHOD__
+        ), E_USER_DEPRECATED);
+
         return new LifecycleEventArgs($document, $documentManager);
     }
 }